requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
sqlalchemy==2.0.23
requests==2.31.0
gunicorn==21.2.0
